pyproject.toml
[project]
name = "marc-reasoning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.73.0",
    "colorcet>=3.1.0",
    "datasets>=4.4.1",
    "google-genai>=1.50.1",
    "google-generativeai>=0.8.5",
    "ipywidgets>=8.1.8",
    "jupyterlab>=4.4.10",
    "matplotlib>=3.10.7",
    "openai>=2.8.0",
    "pandas>=2.3.3",
    "pyperclip>=1.11.0",
    "python-docx>=1.2.0",
    "pyyaml>=6.0.3",
    "scienceplots>=2.1.1",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "selenium>=4.39.0",
    "statsmodels>=0.14.5",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
    "xai-sdk>=1.5.0",
]
